+++ pyengine/Components/SpriteComponent.py
@@ -42,9 +42,8 @@ class SpriteComponent:
     @scale.setter
     def scale(self, scale):
         self.__scale = scale
-        self.width = int(self.width*scale)
-        self.height = int(self.height*scale)
-        self.entity.image = pygame.transform.scale(self.entity.image, (self.width, self.height))
+        self.entity.image = pygame.transform.scale(self.entity.image, (int(self.width * scale),
+                                                                       int(self.height * scale)))
         self.update_entity()
 
     @property